consolidate SubscriptionExecutor into Executor
